requirements.txt
agate==1.7.1
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.9
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
clickhouse-connect==0.6.13
csvkit==1.0.7
dbfread==2.0.7
et-xmlfile==1.1.0
greenlet==2.0.2
idna==3.4
isodate==0.6.1
leather==0.3.4
numpy==1.24.2
olefile==0.46
openpyxl==3.1.2
pandas==1.4.3
parsedatetime==2.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.49
text-unidecode==1.3
urllib3==2.0.4
xlrd==2.0.1
